cobc and gdb on docker cannot take win32 paths as arguments. (#90)
* cobc and gdb on docker cannot take win32 paths as arguments.

When the plugin is used on a windows computer with the docker image running, win32 style paths break cobc and gdb. This fixes cobc, I tried to make it unintrusive by checking if docker and win32 were both used before fixing the path. I could not fix gdb yet, that problem seems to be much more complex because of SourceMap.

* fix docker run on win32

* give gdb posix paths